Refactored services to lib folder
Necessary to start organization our objects within the `lib` folder in support of this application.
